healthcheck/requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-APScheduler==1.13.0
python-dotenv==1.0.0 